+++ pkg/demo/package.json
@@ -13,7 +13,7 @@
     "ethers": "^5.3.0",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
-    "react-spring": "^9.2.1",
+    "react-spring": "^9.2.2",
     "stylis": "^4.0.10",
     "twgl.js": "^4.19.1",
     "uikit": "workspace:pkg/uikit",
@@ -25,6 +25,6 @@
     "@types/react-dom": "^17.0.6",
     "@vitejs/plugin-react-refresh": "^1.3.3",
     "typescript": "^4.3.2",
-    "vite": "^2.3.6"
+    "vite": "^2.3.7"
   }
 }
