+++ pkg/uikit/src/index.tsx
@@ -3,7 +3,7 @@ export { BaseUrl, useBaseUrl } from "./BaseUrl"
 export { Button } from "./Button"
 export { ButtonArea } from "./ButtonArea"
 export { ButtonIcon } from "./ButtonIcon"
-export { GLMoire, Moire, useMoireImage } from "./Moire"
+export { Moire } from "./Moire"
 export { TextInput } from "./TextInput"
 export { Uikit } from "./Uikit"
 export { addSlash, lerp, list, map, shuffle, useUid, raf } from "./utils"
