+++ pkg/website/package.json
@@ -11,23 +11,24 @@
   "dependencies": {
     "@emotion/cache": "^11.4.0",
     "@emotion/react": "^11.4.0",
-    "@mdx-js/mdx": "next",
+    "@mdx-js/mdx": "^2.0.0-next.9",
+    "@mdx-js/preact": "^1.6.22",
     "@mdx-js/react": "^1.6.22",
     "react": "^17.0.2",
-    "react-cool-inview": "^2.0.5",
+    "react-cool-inview": "^2.0.7",
     "react-dom": "^17.0.2",
-    "react-spring": "^9.1.2",
+    "react-spring": "^9.2.2",
     "stylis": "^4.0.10",
     "uikit": "workspace:pkg/uikit",
-    "use-nft": "^0.6.2",
+    "use-nft": "^0.9.0",
     "use-viewport": "git://github.com/bpierre/use-viewport.git#master"
   },
   "devDependencies": {
     "@vitejs/plugin-react-refresh": "^1.3.3",
-    "eslint": "^7.26.0",
-    "eslint-plugin-react": "^7.23.2",
+    "eslint": "^7.28.0",
+    "eslint-plugin-react": "^7.24.0",
     "eslint-plugin-react-hooks": "^4.2.0",
-    "vite": "^2.3.2",
-    "vite-plugin-mdx": "^3.5.5"
+    "vite": "^2.3.7",
+    "vite-plugin-mdx": "^3.5.6"
   }
 }
