+++ pkg/website/src/Home/Steps.jsx
@@ -1,7 +1,7 @@
 import React, { useMemo } from "react"
 import { css } from "@emotion/react"
 import { useSpring, animated, config as springConfig } from "react-spring"
-import { GLMoire, fonts, theme, gu } from "uikit"
+import { Moire, fonts, theme, gu } from "uikit"
 import useInView from "react-cool-inview"
 import { useLayout } from "../styles.js"
 import { steps } from "../content.jsx"
@@ -218,7 +218,7 @@ function VisualMoire() {
         height: ${styles.visualsSize}px;
       `}
     >
-      <GLMoire
+      <Moire
         animate={inView}
         width={styles.visualsSize * 1.5}
         height={styles.visualsSize * 1.5}
